목요일 작업 시작 준비
