+++ src/app/Hotel/page.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import Hotel from "./Hotel";
+
+const HotelView = () => {
+  return <Hotel></Hotel>;
+};
+
+export default HotelView;
