+++ src/app/ProductList/ProductList.module.scss
@@ -1,6 +1,12 @@
 /*헤더*/
 @import url("https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100..900&display=swap");
 
+body {
+  width: 360px;
+  height: 814px;
+  margin: 0 auto;
+}
+
 :root {
   --font: "Noto Sans KR", sans-serif;
 }
@@ -74,6 +80,7 @@
   text-align: left;
   padding: 10px 20px;
 }
+
 .main-container .list-container {
   width: 100%;
   max-width: 360px;
@@ -82,16 +89,43 @@
   justify-content: flex-start;
   gap: 10px;
   padding: 0 20px;
-  overflow-x: hidden;
-  overflow-y: hidden;
+  overflow-x: auto;
+  flex-wrap: nowrap;
 }
+
+.main-container .currentList .list-item {
+  list-style: none;
+
+  flex-direction: column;
+  align-items: center;
+  width: 120px;
+  height: 180px;
+  flex-shrink: 0;
+}
+
 .main-container .currentList .list-image {
   margin-bottom: 10px;
 }
-.main-container .currentList .list-item {
-  list-style: none;
-  display: flex;
-  flex-direction: column;
-  width: 120px;
-  height: 180px;
+
+.main-container .currentList .list-image img {
+  width: 100%;
+  height: 120px;
+  object-fit: cover;
+  border-radius: 10px;
+}
+
+.main-container .currentList .list-title p {
+  font-family: Noto Sans KR;
+  font-size: 14px;
+  font-weight: 500;
+  line-height: 20.27px;
+  letter-spacing: -0.08em;
+}
+
+.main-container .currentList .list-price p {
+  font-family: Noto Sans KR;
+  font-size: 16px;
+  font-weight: 700;
+  line-height: 23.17px;
+  letter-spacing: -0.08em;
 }
